Moved Dashboard tutorial to dashboard
